requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.22
langgraph>=0.0.25
openai>=1.0
httpx>=0.24
Jinja2>=3.1
python-dotenv>=1.0
pytest>=7.2
langsmith==0.4.28